fix: useSearch component typescript and bug fix
